chore: moved hook to a new file and renamed a state
